train resnet

+++ fine_tuning.py
@@ -292,7 +292,7 @@ if __name__ == "__main__":
 		resnet_log.write("ratio:{}|known:{}\n".format(close_ratios[i],','.join(classes)))
 		criterion = nn.CrossEntropyLoss()
 		optimizer_model = optimizer(model)
-		output_model = str(close_ratios[i] * 10)
+		output_model = str(int(close_ratios[i] * 10))
 		train_model(model,dataloaders_dict,criterion,optimizer_model,output_model,num_epochs)
 	resnet_log.close()
 	# test_model(model,"mynet.pkl")
